refactor: Move files to /internals
